# docs/openapi/discodeit-api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "/",
      "description": "배포된 서버"
    }
  ],
...
  "paths": {
...
    "/api/messages": {
      "get": {
        "tags": [
          "Message"
        ],
        "summary": "Channel의 Message 목록 조회",
        "operationId": "findAllByChannelId",
        "parameters": [
          {
            "name": "channelId",
            "in": "query",
            "description": "조회할 Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Message"
        ],
        "summary": "Message 생성",
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "messageCreateRequest"
                ],
                "type": "object",
                "properties": {
                  "messageCreateRequest": {
                    "$ref": "#/components/schemas/MessageCreateRequest"
                  },
                  "attachments": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message가 성공적으로 생성됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          },
          "400": {
            "description": "요청 값이 올바르지 않음"
          },
          "404": {
            "description": "Channel 또는 User를 찾을 수 없음"
          }
        }
      }
    },
...
    "/api/messages/{messageId}": {
      "delete": {
        "tags": [
          "Message"
        ],
        "summary": "Message 삭제",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "삭제할 Message ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Message가 성공적으로 삭제됨"
          },
          "404": {
            "description": "Message를 찾을 수 없음"
          }
        }
      },
      "patch": {
        "tags": [
          "Message"
        ],
        "summary": "Message 내용 수정",
        "operationId": "update_1",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "수정할 Message ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Message가 성공적으로 수정됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          },
          "400": {
            "description": "요청 값이 올바르지 않음"
          },
          "404": {
            "description": "Message를 찾을 수 없음"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageCreateRequest": {
        "required": [
          "authorId",
          "channelId",
          "content"
        ],
        "type": "object",
        "properties": {
          "content": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "MessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "content": {
            "type": "string"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          },
          "attachmentIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "MessageUpdateRequest": {
        "required": [
          "newContent"
        ],
        "type": "object",
        "properties": {
          "newContent": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
    }
  }
}